+++ example/lib/tela-teste-card.dart
@@ -4,13 +4,16 @@ import 'package:z_components/components/z-item-tile-convite.dart';
 import 'package:z_components/components/z-selection/z-selection-item.dart';
 import 'package:z_components/components/z-collection/z-collection-item.dart';
 import 'package:z_components/components/z-collection/z-collection.dart';
-import 'package:z_components/config/z-tipo-header.dart';
-import 'package:z_components/components/z-header.dart';
-import 'package:z_components/components/z-selection/z-selection.dart';
 import 'package:z_components/components/z-inputs/z-input-data-padrao.dart';
-import 'package:z_components/components/z-estrutura-empresa/z-estrutura-empresa.dart';
-import 'package:z_components/components/z-estrutura-empresa/bloc/z-estrutura-empresa-cubit.dart';
 import 'package:z_components/components/z-item-tile-pergunta-adicional.dart';
+import 'package:z_components/components/z-inputs/z-input-data-expiracao.dart';
+import 'package:z_components/components/z-form/presenter/z-form.dart';
+import 'package:z_components/components/z-form/view/z-form-viewmodel.dart';
+import 'package:z_components/components/z-documentos/presenter/lista-documentos.dart';
+import 'package:z_components/view-model/colaborador-documento-viewmodel.dart';
+import 'package:z_components/view-model/documento-campo.dart';
+import 'package:z_components/components/z-endereco/z-input-endereco.dart';
+import 'package:z_components/components/z-tile.dart';
 
 class TelaTesteCard extends StatefulWidget {
   @override
@@ -35,11 +38,84 @@ class _TelaTesteCardState extends State<TelaTesteCard> {
   GlobalKey<ZCollectionState> keyPorte = new GlobalKey<ZCollectionState>();
 
   List<ZCollectionItem> listaPorte = [
-    ZCollectionItem(valor: "T-E-S-T", chave: "TesteChave", titulo: "Teste"),
+    ZCollectionItem(
+        valor:
+            "Instalador de isolantes térmicos de caldeira e tubulações de teste 123 testando overflow",
+        chave:
+            "Instalador de isolantes térmicos de caldeira e tubulações de teste 123 testando overflow",
+        titulo:
+            "Instalador de isolantes térmicos de caldeira e tubulações de teste 123 testando overflow"),
     ZCollectionItem(valor: "Teste2", chave: "Teste2", titulo: "Teste2"),
     ZCollectionItem(valor: "Teste3", chave: "Teste3", titulo: "Teste3"),
     ZCollectionItem(valor: "Teste4", chave: "Teste4", titulo: "Teste4"),
     ZCollectionItem(valor: "Teste5", chave: "Teste5", titulo: "Teste5"),
+  ];
+
+  List<ZFormViewModel> listaForm = [
+    ZFormViewModel(
+        nomeCampo: "Nome Completo",
+        obrigatorio: true,
+        enabled: true,
+        tipoValorCampo: "text",
+        controller: TextEditingController()),
+    ZFormViewModel(
+        nomeCampo: "Data Nascimento",
+        obrigatorio: false,
+        enabled: true,
+        maxLength: 10,
+        tipoValorCampo: "date",
+        controller: TextEditingController()),
+    ZFormViewModel(
+        nomeCampo: "Escolaridade",
+        obrigatorio: false,
+        enabled: true,
+        tipoValorCampo: "text",
+        controller: TextEditingController()),
+    ZFormViewModel(
+        nomeCampo: "Email",
+        obrigatorio: true,
+        enabled: true,
+        tipoValorCampo: "text",
+        controller: TextEditingController()),
+    ZFormViewModel(
+        nomeCampo: "Celular",
+        obrigatorio: true,
+        enabled: true,
+        tipoValorCampo: "celular",
+        controller: TextEditingController()),
+  ];
+
+  List<ColaboradorDocumentoViewModel> _listaDocumentosTemporario = [
+    new ColaboradorDocumentoViewModel(
+        nomeDocumento: "RG",
+        campos: [],
+        status: "",
+        horizontalOuVertical: "",
+        idColaborador: "",
+        idDocumento: "",
+        idImagemDocumento: "",
+        qtdePaginaUpload: "0"
+    ),
+    new ColaboradorDocumentoViewModel(
+        nomeDocumento: "Teste",
+        campos: [],
+        status: "",
+        horizontalOuVertical: "",
+        idColaborador: "",
+        idDocumento: "",
+        idImagemDocumento: "",
+        qtdePaginaUpload: "0"
+    ),
+    new ColaboradorDocumentoViewModel(
+        nomeDocumento: "Outro",
+        campos: [],
+        status: "",
+        horizontalOuVertical: "",
+        idColaborador: "",
+        idDocumento: "",
+        idImagemDocumento: "",
+        qtdePaginaUpload: "0"
+    ),
   ];
 
   @override
@@ -81,14 +157,23 @@ class _TelaTesteCardState extends State<TelaTesteCard> {
     super.initState();
   }
 
+  bool editado = false;
+
   @override
   Widget build(BuildContext context) {
     return new Scaffold(
+      floatingActionButton: FloatingActionButton(
+        onPressed: (){
+          setState(() {
+            editado = !editado;
+          });
+        },
+      ),
       appBar: AppBar(
         leading: new Text("TESTE"),
       ),
       body: new Container(
-        child: new Column(
+        child: new ListView(
           children: [
             new ZInputDataPadrao(
               themeData: Theme.of(context),
@@ -104,17 +189,73 @@ class _TelaTesteCardState extends State<TelaTesteCard> {
               legenda: "Data final",
               validacao: (validacaoDataFinal) {},
             ),
-            new ZSelection(
-              titulo: "Teste",
-              lista: listaSelecao,
-              listaRespostas: listaSelecaoRespostas,
+            new ZInputDataExpiracao(
               themeData: Theme.of(context),
-              onChange: (value) {},
+              controllerData: numero,
+              onChange: (textoDataExpiracao) {},
+              dataFocus: FocusNode(),
+              legenda: "Prazo Preenchimento",
+              validacao: (validacaoDataFinal) {
+                if (validacaoDataFinal) {
+                  print("valido");
+                } else {
+                  print("não valido");
+                }
+              },
             ),
-            ZItemTilePerguntaAdicional(
-              textoPergunta: "Teste",
-              onPressed: () {},
-            )
+            new ZInputEndereco(
+              themeData: Theme.of(context),
+              campoObrigatorio: true,
+              cepController: cep,
+              logradouroController: logradouro,
+              numeroController: numero,
+              complementoController: complemento,
+              bairroController: bairro,
+              cidadeController: cidade,
+              estadoController: estado,
+              validacao: (validarEndereco){},
+            ),
+            new ZTile(
+              onTap: () {
+                Navigator.push(
+                    context,
+                    new MaterialPageRoute(
+                        builder: (context) => new ListaDocumentos(
+                          iconeInformativo: true,
+                          textoInformativo: "Será necessário que nos encaminhe uma foto /imagem dos documentos solicitados assim como  o preenchimento de dados requisitados",
+                          tituloTextoInformativo: "Informação de Documento",
+                          colaboradorDocumentoViewModel: _listaDocumentosTemporario
+                              .map((e) => new ColaboradorDocumentoViewModel(
+                            nomeDocumento: e.nomeDocumento,
+                            campos: e.campos,
+                            idDocumento: e.idDocumento,
+                            idImagemDocumento: e.idDocumento,
+                            qtdePaginaUpload: "0/${e.qtdePaginaUpload}",
+                            imagemObrigatoria: true,
+                          ))
+                              .toList(),
+                          retornarListaDocumentos: (documento) {},
+                        )));
+              },
+              leading: new Row(
+                children: [
+                  new Icon(
+                    Icons.wysiwyg,
+                    color: Colors.grey,
+                  ),
+                  new Container(
+                    margin: const EdgeInsets.only(left: 8.0),
+                    child: new Text("Documentos"),
+                  )
+                ],
+              ),
+              trailing: new Container(
+                child: new Icon(
+                  Icons.arrow_forward_ios_outlined,
+                  color: Colors.grey,
+                ),
+              ),
+            ),
           ],
         ),
       ),