+++ lib/components/z-documentos/presenter/lista-documentos.dart
@@ -1,15 +1,33 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:z_components/components/utils/documento-status.dart';
+import 'package:z_components/components/utils/icone-voltar.dart';
 import 'package:z_components/components/utils/svg.dart';
 import 'package:z_components/components/z-documentos/view/lista-documentos-view.dart';
+import 'package:z_components/config/z-dialog.dart';
+import 'package:z_components/view-model/colaborador-documento-viewmodel.dart';
+
+import '../../z-alert-dialog.dart';
 
 class ListaDocumentos extends StatefulWidget {
   String idColaborador;
   String token;
   String keyGeniusScan;
+  List<ColaboradorDocumentoViewModel> colaboradorDocumentoViewModel;
+  Function(ColaboradorDocumentoViewModel) retornarListaDocumentos;
+  bool iconeInformativo;
+  String textoInformativo;
+  String tituloTextoInformativo;
 
-  ListaDocumentos({this.idColaborador, this.token, this.keyGeniusScan});
+  ListaDocumentos(
+      {this.idColaborador,
+      this.token,
+      this.keyGeniusScan,
+      this.colaboradorDocumentoViewModel,
+      this.retornarListaDocumentos,
+      this.iconeInformativo = false,
+      this.textoInformativo = "",
+      this.tituloTextoInformativo = ""});
 
   @override
   _ListaDocumentosState createState() => _ListaDocumentosState();
@@ -17,7 +35,7 @@ class ListaDocumentos extends StatefulWidget {
 
 class _ListaDocumentosState extends State<ListaDocumentos> {
   ListaDocumentosView _view;
-
+  bool _documentoAtualizado = true;
   @override
   void initState() {
     _view = ListaDocumentosView(this);
@@ -29,11 +47,90 @@ class _ListaDocumentosState extends State<ListaDocumentos> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
+        actions: [
+          _retornarDialogInformativo()
+        ],
+        leading: IconeVoltar(),
         centerTitle: true,
         title: new Text("DOCUMENTOS"),
       ),
       body: _listarDocumentos(),
     );
+  }
+
+  Widget _retornarDialogInformativo(){
+    if(widget.iconeInformativo){
+      return new IconButton(
+        onPressed: (){
+          showDialog(
+              context: context,
+              builder: (context) {
+                return new ZAlertDialog(
+                  zDialog: ZDialog.normal,
+                  colorLine: new Color(0xFF1e26f7),
+                  child: new Column(
+                    children: [
+                      new Container(
+                          margin: EdgeInsets.only(top: 16, bottom: 4),
+                          child: Icon(
+                            Icons.info,
+                            color: Color(0xFF1e26f7),
+                          )),
+                      new Container(
+                        margin: const EdgeInsets.only(
+                            bottom: 8,
+                            left: 16,
+                            right: 16,),
+                        child: new Text(
+                          widget.tituloTextoInformativo,
+                          textAlign: TextAlign.center,
+                          style: new TextStyle(
+                            fontWeight: FontWeight.bold
+                          ),
+                        ),
+                      ),
+                      new Container(
+                        margin: const EdgeInsets.only(
+                            bottom: 8,
+                            left: 16,
+                            right: 16,),
+                        child: new Text(
+                          widget.textoInformativo,
+                          textAlign: TextAlign.center,
+                        ),
+                      ),
+                      new Divider(
+                        height: 1.0,
+                        color: Colors.black,
+                      ),
+                      new Container(
+                        margin: const EdgeInsets.only(
+                            bottom: 8, top: 8),
+                        child: new GestureDetector(
+                          onTap: () {
+                            Navigator.pop(context);
+                          },
+                          child: new Text(
+                              "OK",
+                              style: new TextStyle(
+                                fontWeight: FontWeight.bold
+                              ),
+                          ),
+                        ),
+                      )
+                    ],
+                  ),
+                );
+              });
+        },
+        icon: new Icon(
+          Icons.info,
+          size: 28,
+        ),
+      );
+    }else{
+      return new Container();
+    }
   }
 
   Widget _listarDocumentos() {
@@ -45,11 +142,11 @@ class _ListaDocumentosState extends State<ListaDocumentos> {
           return new GestureDetector(
             onTap: () async {
               await _view.atualizarDocumento(index);
+              setState(() {});
             },
             child: new Container(
               margin: EdgeInsets.only(top: 1),
               height: 50,
-              //margin: EdgeInsets.only(left: 16.0),
               decoration: new BoxDecoration(
                 color: Color(0xffffffff),
               ),
@@ -75,22 +172,45 @@ class _ListaDocumentosState extends State<ListaDocumentos> {
                           item.nomeDocumento,
                           style: new TextStyle(
                             color: Color(0xff000000),
-                            fontWeight: FontWeight.bold,
-                            fontSize: 12,
+                            fontSize: 14,
                           ),
                         ),
                       ),
                     ],
                   ),
-                  new Container(
-                    margin: EdgeInsets.only(right: 16),
-                    child: _buildIconeStatus(item.status.toUpperCase()),
+                  new Row(
+                    children: [
+                      _retornarStatus(index),
+                      new Container(
+                        margin: EdgeInsets.only(right: 16),
+                        child: new Icon(
+                          Icons.chevron_right,
+                          color: Colors.black,
+                        ),
+                      ),
+                    ],
                   )
                 ],
               ),
             ),
           );
         });
+  }
+
+  Widget _retornarStatus(int index){
+    if(_view.listaDocumentos[index].documentoAtualizado){
+      return new Container(
+        decoration: new ShapeDecoration(
+            shape: CircleBorder(), color: Colors.green),
+        child: Icon(
+          Icons.ac_unit,
+          color: Colors.transparent,
+          size: 9,
+        ),
+      );
+    }else{
+      return new Container();
+    }
   }
 
   Widget _buildIconeStatus(String status) {