+++ lib/components/z-form/presenter/z-form.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:z_components/components/z-form/view/z-form-view.dart';
 import 'package:z_components/components/z-form/view/z-form-viewmodel.dart';
-import 'package:z_components/components/z-inputs/z-input-celular.dart';
-import 'package:z_components/components/z-inputs/z-input-cpf.dart';
 import 'package:z_components/components/z-inputs/z-input-generic.dart';
 
 class ZForm extends StatefulWidget {
@@ -38,7 +36,7 @@ class _ZFormState extends State<ZForm> {
 
   List<Widget> _montarListaDinamicaDocumentos() {
     List<Widget> lista = new List();
-    for (int i = 0; i < _view.listaRespostasUsuario.length; i++) {
+    for (int i = 0; i < widget.viewmodel.length; i++) {
       var item = widget.viewmodel[i];
       if (item.tipoValorCampo == "date") {
         lista.add(new Column(
@@ -52,17 +50,14 @@ class _ZFormState extends State<ZForm> {
                 tipoTeclado: TextInputType.number,
                 maxLength: item.maxLength,
                 comMascara: true,
-                inputPadraoFocus: FocusNode(),
                 textMask: "##/##/####",
                 hintText: "DD / MM / AAAA",
-                controllerInputPadrao: TextEditingController(
-                  text: _view.listaRespostasUsuario[i],
-                ),
+                controllerInputPadrao: item.controller,
                 onChange: (text) {
                   if (text.length == 10) {
                     var textoSeparado = text.split("/");
                     item.resposta =
-                    "${textoSeparado[2]}-${textoSeparado[1]}-${textoSeparado[0]}";
+                        "${textoSeparado[2]}-${textoSeparado[1]}-${textoSeparado[0]}";
                     widget.onChange(widget.viewmodel);
                   }
                 },
@@ -84,9 +79,7 @@ class _ZFormState extends State<ZForm> {
                 titulo: item.nomeCampo,
                 tipoTeclado: TextInputType.number,
                 maxLength: item.maxLength,
-                inputPadraoFocus: FocusNode(),
-                controllerInputPadrao:
-                TextEditingController(text: widget.viewmodel[i].resposta),
+                controllerInputPadrao: item.controller,
                 onChange: (text) {
                   item.resposta = text.trim();
                   widget.onChange(widget.viewmodel);
@@ -112,12 +105,13 @@ class _ZFormState extends State<ZForm> {
                 textMask: "(##) #####-####",
                 maxLength: item.maxLength,
                 hintText: "( ** ) 9 **** - ****",
-                inputPadraoFocus: FocusNode(),
-                controllerInputPadrao:
-                TextEditingController(text: widget.viewmodel[i].resposta),
+                controllerInputPadrao: item.controller,
                 onChange: (text) {
-                  item.resposta = text.trim();
-                  widget.onChange(widget.viewmodel);
+                  if (text.length == 15) {
+                    item.resposta = text.trim();
+                    widget.onChange(widget.viewmodel);
+                    item.controller.text = item.resposta;
+                  }
                 },
               ),
             ),
@@ -134,15 +128,13 @@ class _ZFormState extends State<ZForm> {
                 campoObrigatorio: item.obrigatorio,
                 enabled: item.enabled,
                 themeData: Theme.of(context),
-                inputPadraoFocus: FocusNode(),
                 tipoTeclado: TextInputType.number,
                 titulo: item.nomeCampo,
                 maxLength: item.maxLength,
                 comMascara: true,
                 textMask: "###.###.###-##",
                 hintText: "000.000.000-00",
-                controllerInputPadrao:
-                TextEditingController(text: widget.viewmodel[i].resposta),
+                controllerInputPadrao: item.controller,
                 onChange: (text) {
                   item.resposta = text.trim();
                   widget.onChange(widget.viewmodel);
@@ -163,10 +155,8 @@ class _ZFormState extends State<ZForm> {
                 enabled: item.enabled,
                 titulo: item.nomeCampo,
                 themeData: Theme.of(context),
-                inputPadraoFocus: FocusNode(),
                 maxLength: item.maxLength,
-                controllerInputPadrao:
-                TextEditingController(text: widget.viewmodel[i].resposta),
+                controllerInputPadrao: item.controller,
                 onChange: (text) {
                   item.resposta = text.trim();
                   widget.onChange(widget.viewmodel);
@@ -188,10 +178,8 @@ class _ZFormState extends State<ZForm> {
                 titulo: item.nomeCampo,
                 tipoTeclado: TextInputType.text,
                 themeData: Theme.of(context),
-                inputPadraoFocus: FocusNode(),
                 maxLength: item.maxLength,
-                controllerInputPadrao:
-                TextEditingController(text: widget.viewmodel[i].resposta),
+                controllerInputPadrao: item.controller,
                 onChange: (text) {
                   item.resposta = text.trim();
                   widget.onChange(widget.viewmodel);
@@ -212,10 +200,8 @@ class _ZFormState extends State<ZForm> {
                 enabled: item.enabled,
                 themeData: Theme.of(context),
                 titulo: item.nomeCampo,
-                inputPadraoFocus: FocusNode(),
                 maxLength: item.maxLength,
-                controllerInputPadrao:
-                TextEditingController(text: widget.viewmodel[i].resposta),
+                controllerInputPadrao: item.controller,
                 onChange: (text) {
                   item.resposta = text.trim();
                   widget.onChange(widget.viewmodel);